update: project structure
